+++ multiply.c
@@ -11,7 +11,7 @@
 
 /*--------------------------------------------------------------------*/
 
-Matrix_t mat_multiply (Matrix_t mat1, Matrix_t mat2){
+Matrix_t mat_multiply(Matrix_t mat1, Matrix_t mat2){
     if (mat1.n != mat2.m) {
         exit(2);
     }
@@ -23,7 +23,7 @@ Matrix_t mat_multiply (Matrix_t mat1, Matrix_t mat2){
     Matrix_t resultingMatrix;
     resultingMatrix.m = p;
     resultingMatrix.n = q;
-    resultingMatrix.elements  = (double*) malloc (sizeof(double) * p * q);
+    resultingMatrix.elements  = (int32_t*) malloc (sizeof(int32_t) * p * q);
 
     for (int i = 0; i < p; i++){
         for (int j = 0; j < q; j++){
