+++ reading.c
@@ -8,11 +8,11 @@
 #include "reading.h"
 
 /*------------------------------------------------------------*/
-static void setArray(double* array, int row, int col) {
+static void setArray(int32_t* array, int row, int col) {
     for (int i = 0; i < row; i++) {
             char temp;
         for (int j = 0; j < col; j++) {
-            scanf("%lf%c", &array[i * col + j], &temp);
+            scanf("%d%c", &array[i * col + j], &temp);
             if (temp != ' ' && j < col - 1) {
                 exit(2);
             }
@@ -24,8 +24,8 @@ static void setArray(double* array, int row, int col) {
 }
 
 void readMatrices(Matrix_t* matrix_A, Matrix_t* matrix_B, int m_A, int m_B, int n_A, int n_B) {
-    double* a = (double*)malloc(sizeof(double) * m_A * n_A);
-    double* b = (double*)malloc(sizeof(double) * m_B * n_B);
+    int32_t* a = (int32_t*)malloc(sizeof(int32_t) * m_A * n_A);
+    int32_t* b = (int32_t*)malloc(sizeof(int32_t) * m_B * n_B);
     if (a == NULL || b == NULL) {
         exit(2);
     }
